Fix monster int sizes (#37)
* Rebalance monsters

* Fix int sizes

+++ Source/loadsave.cpp
@@ -645,15 +645,15 @@ void LoadMonster(LoadHelper *file, Monster &monster)
 	file->Skip(2); // Skip exp - now calculated from monstdat when the monster dies
 
 	if (monster.isPlayerMinion()) // Don't skip for golems
-		monster.toHit = file->NextLE<uint8_t>();
+		monster.toHit = file->NextLE<uint16_t>();
 	else
 		file->Skip(1); // Skip hit as it's already initialized
-	monster.minDamage = file->NextLE<uint8_t>();
-	monster.maxDamage = file->NextLE<uint8_t>();
+	monster.minDamage = file->NextLE<uint16_t>();
+	monster.maxDamage = file->NextLE<uint16_t>();
 	file->Skip(1); // Skip toHitSpecial as it's already initialized
-	monster.minDamageSpecial = file->NextLE<uint8_t>();
-	monster.maxDamageSpecial = file->NextLE<uint8_t>();
-	monster.armorClass = file->NextLE<uint8_t>();
+	monster.minDamageSpecial = file->NextLE<uint16_t>();
+	monster.maxDamageSpecial = file->NextLE<uint16_t>();
+	monster.armorClass = file->NextLE<uint16_t>();
 	file->Skip(1); // Alignment
 	monster.resistance = file->NextLE<uint16_t>();
 	file->Skip(2); // Alignment
@@ -1430,13 +1430,13 @@ void SaveMonster(SaveHelper *file, Monster &monster)
 	file->Skip(1); // Alignment
 	file->WriteLE<uint16_t>(static_cast<uint16_t>(std::min<unsigned>(std::numeric_limits<uint16_t>::max(), monster.exp(sgGameInitInfo.nDifficulty))));
 
-	file->WriteLE<uint8_t>(static_cast<uint8_t>(std::min<uint16_t>(monster.toHit, std::numeric_limits<uint8_t>::max()))); // For backwards compatibility
-	file->WriteLE<uint8_t>(monster.minDamage);
-	file->WriteLE<uint8_t>(monster.maxDamage);
-	file->WriteLE<uint8_t>(static_cast<uint8_t>(std::min<uint16_t>(monster.toHitSpecial(sgGameInitInfo.nDifficulty), std::numeric_limits<uint8_t>::max()))); // For backwards compatibility
-	file->WriteLE<uint8_t>(monster.minDamageSpecial);
-	file->WriteLE<uint8_t>(monster.maxDamageSpecial);
-	file->WriteLE<uint8_t>(monster.armorClass);
+	file->WriteLE<uint16_t>(static_cast<uint8_t>(std::min<uint16_t>(monster.toHit, std::numeric_limits<uint8_t>::max()))); // For backwards compatibility
+	file->WriteLE<uint16_t>(monster.minDamage);
+	file->WriteLE<uint16_t>(monster.maxDamage);
+	file->WriteLE<uint16_t>(static_cast<uint8_t>(std::min<uint16_t>(monster.toHitSpecial(sgGameInitInfo.nDifficulty), std::numeric_limits<uint8_t>::max()))); // For backwards compatibility
+	file->WriteLE<uint16_t>(monster.minDamageSpecial);
+	file->WriteLE<uint16_t>(monster.maxDamageSpecial);
+	file->WriteLE<uint16_t>(monster.armorClass);
 	file->Skip(1); // Alignment
 	file->WriteLE<uint16_t>(monster.resistance);
 	file->Skip(2); // Alignment
